Create README.md
Eager and driven professional looking for an opportunity as a Data Analyst. Proficient in analysis, and teamwork, with a keen interest in leveraging data for problem-solving and business development. Keen to utilize skills in data analysis, processing, governance, and reporting to enhance data processing efficiency and accuracy, ultimately fostering growth and achievement.